Enhance the game over screens
Check if the player won by seeing if there are any aliens left in the
alien array list.
Color the background green if the player won and put a “You WON”
message.
Color the background red if the player lost - meaning the game ended
when the alien array list was not empty.